Muevo código de generación de reporte fuera del modelo del task, a un módulo aparte
