+++ series_tiempo_ar_api/libs/indexing/report/strings.py
@@ -1,0 +1,9 @@
+#!coding=utf8
+
+INDEXING_REPORT_TEMPLATE = \
+    u"""
+{name}:
+    nuevos: {new},
+    actualizados: {updated},
+    totales: {total},
+"""
